# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Swagger documentation for the OpenCost costmodel HTTP API.",
    "title": "OpenCost API",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "paths": {
...
    "/kapis/costwise.wiztelemetry.io/v1alpha1/cloud/config/delete": {
      "get": {
        "description": "删除指定 integrationKey 的云集成配置。通常受管理员鉴权保护。",
        "tags": [
          "CloudConfig"
        ],
        "summary": "删除云集成配置",
        "parameters": [
          {
            "description": "集成 Key。",
            "name": "integrationKey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/kapis/costwise.wiztelemetry.io/v1alpha1/cloud/config/disable": {
      "get": {
        "description": "停用指定 integrationKey 和 source 的云集成配置。通常受管理员鉴权保护。",
        "tags": [
          "CloudConfig"
        ],
        "summary": "停用云集成配置",
        "parameters": [
          {
            "description": "集成 Key。",
            "name": "integrationKey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "配置来源。",
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/kapis/costwise.wiztelemetry.io/v1alpha1/cloud/config/enable": {
      "get": {
        "description": "启用指定 integrationKey 和 source 的云集成配置。通常受管理员鉴权保护。",
        "tags": [
          "CloudConfig"
        ],
        "summary": "启用云集成配置",
        "parameters": [
          {
            "description": "集成 Key。",
            "name": "integrationKey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "配置来源。",
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/kapis/costwise.wiztelemetry.io/v1alpha1/cloud/config/export": {
      "get": {
        "description": "导出全部或指定 integrationKey 的云集成配置，返回内容会自动脱敏。通常受管理员鉴权保护。",
        "tags": [
          "CloudConfig"
        ],
        "summary": "导出云集成配置",
        "parameters": [
          {
            "description": "指定要导出的集成 Key。",
            "name": "integrationKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}